requirements.txt
rich>=13.0.0
pyyaml>=6.0
requests>=2.28.0
pyTelegramBotAPI>=4.12.0
colorama>=0.4.6
psutil>=5.9.0
netifaces>=0.11.0 